+++ src/__tests__/cli-diagnostics.test.ts
@@ -2,7 +2,7 @@ import { test } from 'vitest';
 import assert from 'node:assert/strict';
 import fs from 'node:fs';
 import path from 'node:path';
-import type { DaemonResponse } from '../daemon/client/daemon-client.ts';
+import type { DaemonResponse } from '../daemon-client/daemon-client.ts';
 import { resolveDaemonPaths } from '../daemon/config.ts';
 import {
   runCliCapture as captureCli,
