+++ src/__tests__/cli-exit-paths.test.ts
@@ -13,7 +13,7 @@ import { runCli } from '../cli.ts';
 import { runWebCommand } from '../cli/commands/web.ts';
 import { installIsolatedCliTestEnv } from './cli-test-env.ts';
 import { resolveDaemonPaths } from '../daemon/config.ts';
-import type { DaemonResponse } from '../daemon/client/daemon-client.ts';
+import type { DaemonResponse } from '../daemon-client/daemon-client.ts';
 
 afterEach(() => {
   vi.clearAllMocks();
