+++ src/daemon-client/__tests__/boundary-fault-transport.test.ts
@@ -9,10 +9,10 @@ import {
   closeLoopbackServer,
   listenOnLoopback,
   type LoopbackServer,
-} from '../../../__tests__/test-utils/loopback.ts';
-import { mkdtempForTestSync } from '../../../__tests__/test-utils/tmp-dir.ts';
-import type { DaemonPaths } from '../../config.ts';
-import type { DaemonRequest } from '../../daemon-request.ts';
+} from '../../__tests__/test-utils/loopback.ts';
+import { mkdtempForTestSync } from '../../__tests__/test-utils/tmp-dir.ts';
+import type { DaemonPaths } from '../../daemon/config.ts';
+import type { DaemonRequest } from '../../daemon/daemon-request.ts';
 import { sendRequest } from '../daemon-client-transport.ts';
 
 const { mockRunCmdSync } = vi.hoisted(() => ({ mockRunCmdSync: vi.fn() }));
