+++ src/daemon-client/__tests__/daemon-client-lifecycle.test.ts
@@ -5,7 +5,7 @@ import http from 'node:http';
 import net from 'node:net';
 import path from 'node:path';
 import { afterEach, test, vi } from 'vitest';
-import { mkdtempForTestSync } from '../../../__tests__/test-utils/tmp-dir.ts';
+import { mkdtempForTestSync } from '../../__tests__/test-utils/tmp-dir.ts';
 
 vi.mock('@agent-device/host-kit/command', async (importOriginal) => ({
   ...(await importOriginal<typeof import('@agent-device/host-kit/command')>()),
@@ -18,7 +18,7 @@ vi.mock('@agent-device/host-kit/retry', async (importOriginal) => ({
   sleep: vi.fn(async () => {}),
 }));
 
-import { resolveDaemonPaths, type DaemonPaths } from '../../config.ts';
+import { resolveDaemonPaths, type DaemonPaths } from '../../daemon/config.ts';
 import { sendToDaemon, type DaemonRequest, type DaemonResponse } from '../daemon-client.ts';
 import { attachActiveSessionAddressHint } from '../daemon-client-lifecycle.ts';
 import { computeDaemonCodeSignature } from '@agent-device/host-kit/code-signature';
@@ -27,7 +27,7 @@ import {
   closeLoopbackServer,
   listenOnLoopback,
   supportsLoopbackBind,
-} from '../../../__tests__/test-utils/loopback.ts';
+} from '../../__tests__/test-utils/loopback.ts';
 import { AppError } from '@agent-device/kernel/errors';
 import {
   runCmdDetachedMonitored,
