+++ src/daemon-client/daemon-client-progress.ts
@@ -4,13 +4,13 @@ import type { RequestProgressSink } from '@agent-device/contracts/progress';
 import type http from 'node:http';
 import type { Socket } from 'node:net';
 import { AppError } from '@agent-device/kernel/errors';
-import type { DaemonRequest, DaemonResponse } from '../daemon-request.ts';
+import type { DaemonRequest, DaemonResponse } from '../daemon/daemon-request.ts';
 import { consumeTextLines } from '@agent-device/host-kit/transport';
 import {
   isDaemonProgressEnvelope,
   isDaemonResponseEnvelope,
   shouldStreamRequestProgress,
-} from '../request-progress-protocol.ts';
+} from '../daemon/request-progress-protocol.ts';
 
 type ProgressLineReader = {
   handleLine(line: string): boolean;
