+++ src/daemon-client/daemon-client-rpc.ts
@@ -6,15 +6,15 @@ import {
 } from '@agent-device/kernel/errors';
 import { sanitizeErrorCause } from '@agent-device/kernel/redaction';
 import { createRequestId } from '@agent-device/host-kit/diagnostics';
-import type { DaemonRequest, DaemonResponse } from '../daemon-request.ts';
-import { materializeRemoteArtifacts } from '../../remote/daemon-artifacts.ts';
-import { localizeRemoteDaemonError } from '../../remote/remote-request-diagnostics.ts';
+import type { DaemonRequest, DaemonResponse } from '../daemon/daemon-request.ts';
+import { materializeRemoteArtifacts } from '../remote/daemon-artifacts.ts';
+import { localizeRemoteDaemonError } from '../remote/remote-request-diagnostics.ts';
 import type { DaemonInfo } from './daemon-client-metadata.ts';
 import {
   leaseScopeFromRequest,
   leaseScopeToLeaseRpcParams,
   type LeaseRpcCommand,
-} from '../../core/lease-scope.ts';
+} from '../core/lease-scope.ts';
 
 export function handleDaemonHttpResponseBody(
   body: string,
