+++ src/daemon-client/daemon-client-transport.ts
@@ -2,9 +2,9 @@ import type { RequestProgressSink } from '@agent-device/contracts/progress';
 import net from 'node:net';
 import { AppError } from '@agent-device/kernel/errors';
 import { loadNodeHttpRequester, readNodeHttpResponseBody } from '@agent-device/host-kit/transport';
-import type { DaemonRequest, DaemonResponse } from '../daemon-request.ts';
+import type { DaemonRequest, DaemonResponse } from '../daemon/daemon-request.ts';
 import { emitDiagnostic } from '@agent-device/host-kit/diagnostics';
-import type { DaemonPaths, DaemonTransportPreference } from '../config.ts';
+import type { DaemonPaths, DaemonTransportPreference } from '../daemon/config.ts';
 import {
   readDaemonHttpProgressResponse,
   readDaemonSocketProgressResponse,
