+++ src/daemon-client/daemon-launch-spec.ts
