+++ test/integration/android-emulator-e2e/live-observability-scenario.ts
@@ -3,7 +3,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 
 import { PUBLIC_COMMANDS } from '@agent-device/command-registry/catalog';
-import { readDaemonInfo } from '../../../src/daemon/client/daemon-client-metadata.ts';
+import { readDaemonInfo } from '../../../src/daemon-client/daemon-client-metadata.ts';
 import { resolveDaemonPaths } from '../../../src/daemon/config.ts';
 import {
   collectPagedEventTimeline,
