+++ test/integration/ios-simulator-e2e/live-runner.ts
@@ -5,7 +5,7 @@ import { pathToFileURL } from 'node:url';
 
 import type { AgentDeviceDaemonTransport } from '@agent-device/contracts/client';
 import { PUBLIC_COMMANDS } from '@agent-device/command-registry/catalog';
-import { sendToDaemon } from '../../../src/daemon/client/daemon-client.ts';
+import { sendToDaemon } from '../../../src/daemon-client/daemon-client.ts';
 import { assertPngFile } from '../provider-scenarios/assertions.ts';
 import {
   assertFilesDiffer,
